Cookie Based Login Security - Finished
Cookie is created and counts down from five unsuccessful attempts.
After 5 times it locks login button and sends an email to the
corresponding username trying to be accessed.
As of now it is hard coded without an actual database and uses cookies
rather than sessions.
Will be updated with database and sessions in following weeks.