Login and Register updated
Registration now checks to see if a username is present and does not
allow for addition of same username to the database. Also fixed issue
with phone number - SQL DB had int rather than string for storing of
the phone number, which as an int can exceed the 2billion+ value
limitation.